pyproject.toml
[project]
name = "tapo-cam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "load-dotenv>=0.1.0",
    "opencv-python>=4.12.0.88",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
